fix(connectors): spawn npm through cmd.exe on Windows (shared cmd helper)
After the isPath fix let the seed child take the seeding branch on Windows, `npm
install` failed with empty output: `npm` is `npm.cmd`, and modern Node refuses to
spawn a `.cmd` without a shell (CVE-2024-27980). This was a pre-existing bug in the
`npm()` helper (any `cotal ext add` on Windows) never caught because ext:live is not
in the Windows gate — the connector seed smoke is the first to run `ext add` there.

The manager already solved safe `.cmd`/`.bat` launching (an `append_bat_arg` port).
Lift those pure primitives (quoteCmdArg / buildCmdCommandLine / resolveComspec) from
`implementations/manager/src/runtime/windows-launch.ts` into `@cotal-ai/workspace`
(win-cmd.ts) so BOTH the manager's PtyRuntime launch and the CLI's npm invocation
share one tested contract; the manager re-exports them for its existing smoke. New
`cmdSpawnSpec(command, args)` is the spawnSync-shaped adapter: POSIX passthrough; on
Windows a `.cmd`/`.bat` runs through the system cmd.exe with a byte-for-byte verbatim
command line. `npm()` uses it.

windows-launch smoke passes (moved functions unit-tested off-Windows); seed 47/47,
ext:live 63/63, manager + cli typecheck clean.
